Collection: Sort the list using the parents of reply chains
The solves the issue that certain reply chains were not displayed as one
chain, a result coming from the fact that we
only compared the individual posts.

By comparing the parent of chains instead this is sorted as wanted.

However, while this system works, we might want to improve it later on.
There is probably a lot to optimize there.