Implement reblog/favourite on Post
Now matches the discussion, where there's a function on the Post
that passes itself to the Session to make the calls. The session then
loads the returned JSON and updates the post (on Mastodon).

All values are still GLib object properties, so we can notify on them
to update after changes.

Also extracted a `get_interaction_data` function so that we're not
implementing the same "extract from JSON" code in multiple places.
Other fields shouldn't change between retrieves
(until we support edits!)